requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sse-starlette==2.1.0
pydantic==2.7.4; python_version < "3.13"
pydantic==2.11.7; python_version >= "3.13"
python-dotenv==1.0.1
httpx==0.27.2
requests==2.32.3
feedparser==6.0.11
beautifulsoup4==4.12.3
transformers==4.44.2; python_version < "3.13"
transformers==4.46.0; python_version >= "3.13"
torch==2.4.0; python_version < "3.13"
torch==2.10.0; python_version >= "3.13"
sentence-transformers==3.1.1
spacy==3.7.5; python_version < "3.13"
spacy==3.8.14; python_version >= "3.13"
nltk==3.9.1
scikit-learn==1.5.2